+++ libnd4j/include/loops/broadcasting.h
@@ -112,12 +112,13 @@ class Broadcast {
 
   static void execInverse(int opNum, const void *x, const sd::LongType *xShapeInfo, const void *y,
                           const sd::LongType *yShapeInfo, void *result, const sd::LongType *resultShapeInfo,
-                          int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
+                          long long int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
                           const sd::LongType *tadOffset, const sd::LongType *tadShapeInfoZ,
                           const sd::LongType *tadOffsetZ, uint64_t start, uint64_t stop);
 
   static void exec(int opNum, const void *x, const sd::LongType *xShapeInfo, const void *y,
-                   const sd::LongType *yShapeInfo, void *result, const sd::LongType *resultShapeInfo, int *dimension,
+                   const sd::LongType *yShapeInfo, void *result, const sd::LongType *resultShapeInfo,
+                   long long int *dimension,
                    int dimensionLength, const sd::LongType *tadShapeInfo, const sd::LongType *tadOffset,
                    const sd::LongType *tadShapeInfoZ, const sd::LongType *tadOffsetZ, sd::LoopKind::Kind loopKind,
                    uint64_t start, uint64_t stop);
@@ -135,13 +136,13 @@ class Broadcast {
    */
   template <typename OpType>
   static void exec(const void *x, const sd::LongType *xShapeInfo, const void *y, const sd::LongType *yShapeInfo,
-                   void *result, const sd::LongType *resultShapeInfo, int *dimension, int dimensionLength,
+                   void *result, const sd::LongType *resultShapeInfo, long long int *dimension, int dimensionLength,
                    const sd::LongType *tadShapeInfo, const sd::LongType *tadOffset, const sd::LongType *tadShapeInfoZ,
                    const sd::LongType *tadOffsetZ, sd::LoopKind::Kind loopKind, uint64_t start, uint64_t stop);
 
   template <typename OpType>
   static void execInverse(const void *x, const sd::LongType *xShapeInfo, const void *y, const sd::LongType *yShapeInfo,
-                          void *result, const sd::LongType *resultShapeInfo, int *dimension, int dimensionLength,
+                          void *result, const sd::LongType *resultShapeInfo, long long int *dimension, int dimensionLength,
                           const sd::LongType *tadShapeInfo, const sd::LongType *tadOffset,
                           const sd::LongType *tadShapeInfoZ, const sd::LongType *tadOffsetZ, uint64_t start,
                           uint64_t stop);
