+++ libnd4j/include/loops/reduce_bool.h
@@ -106,7 +106,7 @@ class SD_LIB_HIDDEN ReduceBoolFunction {
                          const sd::LongType *zShapeInfo);
 
   static void exec(int opNum, sd::memory::Workspace *workspace, const void *vx, const sd::LongType *xShapeInfo,
-                   void *vextraParams, void *vz, const sd::LongType *zShapeInfo, const int *dims);
+                   void *vextraParams, void *vz, const sd::LongType *zShapeInfo, const long long int *dims);
 
   /**
    * Execute on the cpu
@@ -122,7 +122,7 @@ class SD_LIB_HIDDEN ReduceBoolFunction {
 
   template <typename OpType>
   static void SD_HOST exec(sd::memory::Workspace *workspace, const void *vx, const sd::LongType *xShapeInfo,
-                           void *vextraParams, void *vz, const sd::LongType *zShapeInfo, const int *dims);
+                           void *vextraParams, void *vz, const sd::LongType *zShapeInfo, const long long int *dims);
 
   /**
    * CPU implementation
