+++ libnd4j/include/ops/declarable/helpers/cpu/percentile.cpp
@@ -29,7 +29,7 @@ namespace helpers {
 
 //////////////////////////////////////////////////////////////////////////
 template <typename T>
-static void _percentile(const NDArray& input, NDArray& output, std::vector<int>& axises, const float q,
+static void _percentile(const NDArray& input, NDArray& output, std::vector<LongType>& axises, const float q,
                         const int interpolation) {
   const int inputRank = input.rankOf();
 
@@ -72,13 +72,13 @@ static void _percentile(const NDArray& input, NDArray& output, std::vector<int>&
   }
 }
 
-void percentile(sd::LaunchContext* context, const NDArray& input, NDArray& output, std::vector<int>& axises,
+void percentile(sd::LaunchContext* context, const NDArray& input, NDArray& output, std::vector<LongType>& axises,
                 const float q, const int interpolation) {
   BUILD_SINGLE_SELECTOR(input.dataType(), _percentile, (input, output, axises, q, interpolation), SD_COMMON_TYPES);
 }
 
 BUILD_SINGLE_TEMPLATE(template void _percentile,
-                      (const NDArray& input, NDArray& output, std::vector<int>& axises, const float q,
+                      (const NDArray& input, NDArray& output, std::vector<LongType>& axises, const float q,
                        const int interpolation),
                       SD_COMMON_TYPES);
